start the Asuit sheet again with problem B second one

+++ src/C_Simple_Calculator.java
@@ -4,9 +4,9 @@ public class C_Simple_Calculator {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         int x = scanner.nextInt();
-        long y = scanner.nextLong();
+        int y = scanner.nextInt();
         System.out.println(x + " + " + y +" = "+ (x+y));
-        System.out.println(x + " * " + y +" = "+ (x*y));
+        System.out.println(x + " * " + y +" = "+ ((long)x*y));
         System.out.println(x + " - " + y +" = "+ (x-y));
     }
 }
